efeed cycling completed.
Tested with a 1 day cycle and it worked. Had tested willpower using the same cycle of 7 days and it worked no reason why this wouldn't work as well.

TODO:
Add checks to efeed to prevent "overfeeding" where the character could specify a higher amount than possible character's bloodpool. This will help auto-manage a few spends as well.

Total sums of emergency table like the nightly job when the efeed is called blank so that an accurate number is given to the player.